+++ transaction-history/display-transactions/displayTransactions.test.js
@@ -9,15 +9,18 @@ describe('show', () => {
     describe('in Rwanda English', () => {
         beforeEach(() => {
             project.vars.lang ='en';
+            //Date.toISOString = jest.fn();
+            //singleRepayment.RepaymentDate = '11/30/2020 12:00:00 AM';
         });
         it('should show a given transaction with sayText in English', () => {
             show(singleRepayment);
             expect(sayText).toHaveBeenCalledWith(
                 `Payment ID: ${singleRepayment.RepaymentId}`
-                +'\nDate Received: 20-05-2020'
+                +'\nDate Received: 2020-02-04'
                 +`\nSeason: ${singleRepayment.Season}`
                 +`\nAmount: ${singleRepayment.Amount} RwF`
                 +`\nPaid from: ${singleRepayment.PaidFrom}`
+                +'\n0) Back'
             );
         });     
         it('should call promptDigits with the given inputhadnler', () => {
@@ -33,10 +36,11 @@ describe('show', () => {
             show(singleRepayment);
             expect(sayText).toHaveBeenCalledWith(
                 `Nomero ndangagikorwa: ${singleRepayment.RepaymentId}`
-                +'\nItariki yishyuriweho: 20-05-2020'
+                +'\nItariki yishyuriweho: 2020-02-04'
                 +`\nIgihembwe: ${singleRepayment.Season}`
                 +`\nAmafaranga: ${singleRepayment.Amount} RwF`
                 +`\nYishyuwe avuye: ${singleRepayment.PaidFrom}`
+                +'\n0) Ahabanza'
             );
         });        
     });
@@ -48,10 +52,11 @@ describe('show', () => {
             show(singleRepayment);
             expect(sayText).toHaveBeenCalledWith(
                 `Rekodi ya malipo: ${singleRepayment.RepaymentId}`
-                +'\nTarehe ya malipo: 20-05-2020'
+                +'\nTarehe ya malipo: 2020-02-04'
                 +`\nMsimu: ${singleRepayment.Season}`
                 +`\nIdadi ya malipo: ${singleRepayment.Amount} KES`
                 +`\nMalipo kutoka simu nambari: ${singleRepayment.PaidFrom}`
+                +'\n0) Rudi nyuma'
             );
         });        
     });
@@ -69,22 +74,22 @@ describe('list', () => {
         it('should list the first four if not given a page', () => {
             list(repaymentsList);
             expect(sayText).toHaveBeenLastCalledWith(
-                'Select a payment for details:'
-                +`\n1. 20-05-2020 - ${repaymentsList[0].Amount} RwF`
-                +`\n2. 05-05-2020 - ${repaymentsList[1].Amount} RwF`
-                +`\n3. 04-03-2020 - ${repaymentsList[2].Amount} RwF`
-                +`\n4. 11-02-2020 - ${repaymentsList[3].Amount} RwF`
+                'Select a payment for details:\n44. Back'
+                +`\n1. 2020-02-04 - ${repaymentsList[0].Amount} RwF`
+                +`\n2. 2020-05-05 - ${repaymentsList[1].Amount} RwF`
+                +`\n3. 2020-04-03 - ${repaymentsList[2].Amount} RwF`
+                +`\n4. 2020-01-01 - ${repaymentsList[3].Amount} RwF`
                 +'\n99. Continue'
             );
         });
         it('should list the second four if given a page number of 2 ', () => {
             list(repaymentsList,2);
             expect(sayText).toHaveBeenLastCalledWith(
-                'Select a payment for details:'
-                +`\n5. 05-02-2020 - ${repaymentsList[4].Amount} RwF`
-                +`\n6. 16-08-2019 - ${repaymentsList[5].Amount} RwF`
-                +`\n7. 05-06-2019 - ${repaymentsList[6].Amount} RwF`
-                +`\n8. 06-05-2019 - ${repaymentsList[7].Amount} RwF`
+                'Select a payment for details:\n44. Back'
+                +`\n5. 2020-05-02 - ${repaymentsList[4].Amount} RwF`
+                +`\n6. 2019-01-17 - ${repaymentsList[5].Amount} RwF`
+                +`\n7. 2019-05-06 - ${repaymentsList[6].Amount} RwF`
+                +`\n8. 2019-06-05 - ${repaymentsList[7].Amount} RwF`
                 +'\n99. Continue'
             );        
         });
@@ -92,11 +97,11 @@ describe('list', () => {
             const erorMessage = 'Something is wrong\n';
             list(repaymentsList,1, erorMessage);
             expect(sayText).toHaveBeenLastCalledWith(
-                erorMessage +'Select a payment for details:'
-                +`\n1. 20-05-2020 - ${repaymentsList[0].Amount} RwF`
-                +`\n2. 05-05-2020 - ${repaymentsList[1].Amount} RwF`
-                +`\n3. 04-03-2020 - ${repaymentsList[2].Amount} RwF`
-                +`\n4. 11-02-2020 - ${repaymentsList[3].Amount} RwF`
+                erorMessage +'Select a payment for details:\n44. Back'
+                +`\n1. 2020-02-04 - ${repaymentsList[0].Amount} RwF`
+                +`\n2. 2020-05-05 - ${repaymentsList[1].Amount} RwF`
+                +`\n3. 2020-04-03 - ${repaymentsList[2].Amount} RwF`
+                +`\n4. 2020-01-01 - ${repaymentsList[3].Amount} RwF`
                 +'\n99. Continue'
             );
         });
@@ -109,22 +114,22 @@ describe('list', () => {
         it('should list the first four if not given a page', () => {
             list(repaymentsList);
             expect(sayText).toHaveBeenLastCalledWith(
-                'Hitamo ubwishyu uhabwe ubusobanuro burambuye:'
-                +`\n1. 20-05-2020 - F${repaymentsList[0].Amount}`
-                +`\n2. 05-05-2020 - F${repaymentsList[1].Amount}`
-                +`\n3. 04-03-2020 - F${repaymentsList[2].Amount}`
-                +`\n4. 11-02-2020 - F${repaymentsList[3].Amount}`
+                'Hitamo ubwishyu uhabwe ubusobanuro burambuye:\n44. Subira Inyuma'
+                +`\n1. 2020-02-04 - F${repaymentsList[0].Amount}`
+                +`\n2. 2020-05-05 - F${repaymentsList[1].Amount}`
+                +`\n3. 2020-04-03 - F${repaymentsList[2].Amount}`
+                +`\n4. 2020-01-01 - F${repaymentsList[3].Amount}`
                 +'\n99. Komeza'
             );
         });
         it('should list the second four if given a page number of 2 ', () => {
             list(repaymentsList,2);
             expect(sayText).toHaveBeenLastCalledWith(
-                'Hitamo ubwishyu uhabwe ubusobanuro burambuye:'
-                +`\n5. 05-02-2020 - F${repaymentsList[4].Amount}`
-                +`\n6. 16-08-2019 - F${repaymentsList[5].Amount}`
-                +`\n7. 05-06-2019 - F${repaymentsList[6].Amount}`
-                +`\n8. 06-05-2019 - F${repaymentsList[7].Amount}`
+                'Hitamo ubwishyu uhabwe ubusobanuro burambuye:\n44. Subira Inyuma'
+                +`\n5. 2020-05-02 - F${repaymentsList[4].Amount}`
+                +`\n6. 2019-01-17 - F${repaymentsList[5].Amount}`
+                +`\n7. 2019-05-06 - F${repaymentsList[6].Amount}`
+                +`\n8. 2019-06-05 - F${repaymentsList[7].Amount}`
                 +'\n99. Komeza'
             );        
         });
@@ -138,22 +143,22 @@ describe('list', () => {
         it('should list the first four if not given a page', () => {
             list(repaymentsList);
             expect(sayText).toHaveBeenLastCalledWith(
-                'Chagua rekodi ya malipo kwa maelezo zaidi:'
-                +`\n1. 20-05-2020 - KES ${repaymentsList[0].Amount}`
-                +`\n2. 05-05-2020 - KES ${repaymentsList[1].Amount}`
-                +`\n3. 04-03-2020 - KES ${repaymentsList[2].Amount}`
-                +`\n4. 11-02-2020 - KES ${repaymentsList[3].Amount}`
+                'Chagua rekodi ya malipo kwa maelezo zaidi:\n44. Rudi nyuma'
+                +`\n1. 2020-02-04 - KES ${repaymentsList[0].Amount}`
+                +`\n2. 2020-05-05 - KES ${repaymentsList[1].Amount}`
+                +`\n3. 2020-04-03 - KES ${repaymentsList[2].Amount}`
+                +`\n4. 2020-01-01 - KES ${repaymentsList[3].Amount}`
                 +'\n99. Endelea'
             );
         });
         it('should list the second four if given a page number of 2 ', () => {
             list(repaymentsList,2);
             expect(sayText).toHaveBeenLastCalledWith(
-                'Chagua rekodi ya malipo kwa maelezo zaidi:'
-                +`\n5. 05-02-2020 - KES ${repaymentsList[4].Amount}`
-                +`\n6. 16-08-2019 - KES ${repaymentsList[5].Amount}`
-                +`\n7. 05-06-2019 - KES ${repaymentsList[6].Amount}`
-                +`\n8. 06-05-2019 - KES ${repaymentsList[7].Amount}`
+                'Chagua rekodi ya malipo kwa maelezo zaidi:\n44. Rudi nyuma'
+                +`\n5. 2020-05-02 - KES ${repaymentsList[4].Amount}`
+                +`\n6. 2019-01-17 - KES ${repaymentsList[5].Amount}`
+                +`\n7. 2019-05-06 - KES ${repaymentsList[6].Amount}`
+                +`\n8. 2019-06-05 - KES ${repaymentsList[7].Amount}`
                 +'\n99. Endelea'
             );        
         });
@@ -163,7 +168,7 @@ describe('list', () => {
 
 const singleRepayment = {
     'RepaymentId': '2534504906',
-    'RepaymentDate': '20/05/2020 20:27:03',
+    'RepaymentDate': '2/4/2020 8:27:03 AM',
     'Season': '2020',
     'Amount': 2001.0000,
     'PaidFrom': '0787117739'
@@ -172,84 +177,84 @@ const repaymentsList = [
     singleRepayment,
     {
         'RepaymentId': '2474515444',
-        'RepaymentDate': '05/05/2020 16:16:56',
+        'RepaymentDate': '5/5/2020 6:16:56 AM',
         'Season': '2020',
         'Amount': 2002.0000,
         'PaidFrom': '0787428878'
     },
     {
         'RepaymentId': '2285699969',
-        'RepaymentDate': '04/03/2020 20:29:54',
+        'RepaymentDate': '4/3/2020 2:29:54 AM',
         'Season': '2020',
         'Amount': 2003.0000,
         'PaidFrom': '663565'
     },
     {
         'RepaymentId': '2233627731',
-        'RepaymentDate': '11/02/2020 21:11:39',
+        'RepaymentDate': '1/1/2020 8:11:39 AM',
         'Season': '2020',
         'Amount': 2004.0000,
         'PaidFrom': '788926'
     },
     {
         'RepaymentId': '2219628297',
-        'RepaymentDate': '05/02/2020 19:36:25',
+        'RepaymentDate': '5/2/2020 8:36:25 AM',
         'Season': '2020',
         'Amount': 2005.0000,
         'PaidFrom': '594206'
     },
     {
         'RepaymentId': '1877459374',
-        'RepaymentDate': '16/08/2019 20:56:53',
+        'RepaymentDate': '1/17/2019 8:56:53 AM',
         'Season': '2020',
         'Amount': 2006.0000,
         'PaidFrom': '0787784673'
     },
     {
         'RepaymentId': '1736415567',
-        'RepaymentDate': '05/06/2019 08:54:32',
+        'RepaymentDate': '5/6/2019 8:54:32 AM',
         'Season': '2019',
         'Amount': 2007.0000,
         'PaidFrom': '0786394610'
     },
     {
         'RepaymentId': '1672889041',
-        'RepaymentDate': '06/05/2019 22:26:21',
+        'RepaymentDate': '6/5/2019 2:26:21 AM',
         'Season': '2019',
         'Amount': 2008.0000,
         'PaidFrom': '0787746931'
     },
     {
         'RepaymentId': '1567436959',
-        'RepaymentDate': '17/03/2019 22:50:29',
+        'RepaymentDate': '1/3/2019 2:50:29 AM',
         'Season': '2019',
         'Amount': 2009.0000,
         'PaidFrom': '0787418369'
     },
     {
         'RepaymentId': '1496045072',
-        'RepaymentDate': '10/02/2019 18:05:58',
+        'RepaymentDate': '1/2/2019 8:05:58 AM',
         'Season': '2019',
         'Amount': 2010.0000,
         'PaidFrom': '0782321660'
     },
     {
         'RepaymentId': '8483197187',
-        'RepaymentDate': '10/01/2019 18:27:50',
+        'RepaymentDate': '1/1/2019 8:27:50 AM',
         'Season': '2019',
         'Amount': 2011.0000,
         'PaidFrom': '0722238622'
     },
     {
         'RepaymentId': '1262988984',
-        'RepaymentDate': '07/10/2018 21:01:53',
+        'RepaymentDate': '7/1/2018 1:01:53 AM',
         'Season': '2019',
         'Amount': 2012.0000,
         'PaidFrom': '225966'
     },
     {
         'RepaymentId': '1144554663',
-        'RepaymentDate': '30/07/2018 21:55:18',
+        'RepaymentDate': '3/7/2018 1:55:18 AM',
         'Season': '2019',
         'Amount': 2013.0000,
         'PaidFrom': '112658'